+++ kauldron/checkpoints/__init__.py
@@ -15,7 +15,7 @@
 """Checkpoints API."""
 
 # pylint: disable=g-importing-member
-
+from kauldron.checkpoints.checkpointer import BaseCheckpointer
 from kauldron.checkpoints.checkpointer import Checkpointer
 from kauldron.checkpoints.checkpointer import NoopCheckpointer
 from kauldron.checkpoints.partial_loader import AbstractPartialLoader
@@ -23,3 +23,4 @@ from kauldron.checkpoints.partial_loader import CkptSource
 from kauldron.checkpoints.partial_loader import KauldronSource
 from kauldron.checkpoints.partial_loader import PartialLoader
 from kauldron.checkpoints.pytree_checkpoint import PyTreeCheckpointer
+# pylint: enable=g-importing-member