+++ kauldron/checkpoints/partial_loader_test.py
@@ -74,22 +74,15 @@ def old_trainer(tmp_path: pathlib.Path):
   return cfg
 
 
-def _make_chpt(
+def _make_loader(
     old_trainer: kd.train.Trainer,  # pylint: disable=redefined-outer-name
-    new_trainer: kd.train.Trainer,  # pylint: disable=redefined-outer-name
     new_to_old: dict[str, str],
 ):
   loader = kd.ckpts.PartialLoader(  # pytype: disable=wrong-arg-types
       source=kd.ckpts.KauldronSource(old_trainer.workdir),  # pylint: disable=missing-kwoa
       new_to_old=new_to_old,
   )
-
-  ckpt = kd.ckpts.Checkpointer(  # pytype: disable=wrong-arg-types
-      workdir=new_trainer.workdir,
-      save_interval_steps=1,
-      partial_initializer=loader,
-  )
-  return ckpt
+  return loader
 
 
 def test_loader(new_trainer: kd.train.Trainer, old_trainer: kd.train.Trainer):  # pylint: disable=redefined-outer-name
@@ -98,14 +91,13 @@ def test_loader(new_trainer: kd.train.Trainer, old_trainer: kd.train.Trainer):  
 
   init_state = new_trainer.init_state()
 
-  ckpt = _make_chpt(
+  loader = _make_loader(
       old_trainer,
-      new_trainer,
       new_to_old={
           'params.decoder.layers_1': 'params.encoder',
       },
   )
-  new_state = ckpt.restore(init_state)
+  new_state = loader.transform(init_state)
   assert_utils.assert_trees_all_same_type(init_state, new_state)
 
   # New weights have been copied
@@ -128,14 +120,13 @@ def test_loader(new_trainer: kd.train.Trainer, old_trainer: kd.train.Trainer):  
       init_state.params['decoder']['layers_0'],
   )
 
-  ckpt = _make_chpt(
+  loader = _make_loader(
       old_trainer,
-      new_trainer,
       new_to_old={
           'params.encoder': 'params.encoder',
       },
   )
-  new_state = ckpt.restore(init_state)
+  new_state = loader.transform(init_state)
   chex.assert_trees_all_close(
       new_state.params['encoder'],
       old_state.params['encoder'],